Fix the energy loss issue between MJ and RF Caused by loss of precision Changed default values in configuration for Thermal Expansion to 437.5

+++ src/minecraft/powercrystals/powerconverters/common/TileEntityEnergyBridge.java
@@ -73,7 +73,7 @@ public class TileEntityEnergyBridge extends TileEntity implements INeighboorUpda
                     if (energyNotProduced > energyRemaining) {
                         energyNotProduced = energyRemaining;
                     }
-                    _producerOutputRates.put(prod.getKey(), (energyRemaining - energyNotProduced) / prod.getValue().getPowerSystem().getInternalEnergyPerOutput());
+                    _producerOutputRates.put(prod.getKey(), (int) ((energyRemaining - energyNotProduced) / prod.getValue().getPowerSystem().getInternalEnergyPerOutput()));
                     energyRemaining = energyNotProduced;
                 } else {
                     prod.getValue().produceEnergy(0);
